+++ examples/RSC/app/(dashboard)/settings/user-info.server.tsx
@@ -3,11 +3,11 @@ import ClientUserInfo from "./user-info.client"
 import { layoutServerStorage } from "@/context/layout.storage.server"
 
 export default async function ServerUserInfo() {
-  const name = pageServerStorage.user.name.getState()
-  const role = pageServerStorage.user.role.getState()
-  const city = pageServerStorage.person.address.city.getState()
+  const name = await pageServerStorage.user.name.getState()
+  const role = await pageServerStorage.user.role.getState()
+  const city = await pageServerStorage.person.address.city.getState()
 
-  const session = layoutServerStorage.session.getState()
+  const session = await layoutServerStorage.session.getState()
 
   return (
     <>